requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
httpx==0.28.1
python-multipart==0.0.20
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
